+++ src/main/ruby/truffleruby/core/symbol.rb
@@ -29,10 +29,6 @@
 class Symbol
   include Comparable
 
-  def self.all_symbols
-    ObjectSpace.each_object(Symbol).to_a
-  end
-
   def <=>(other)
     return unless other.kind_of? Symbol